Move Git and worktree ownership into package
