+++ apps/server/src/workspace/layers/file-search.ts
@@ -6,7 +6,7 @@ import fuzzysort from "fuzzysort";
 
 import { FsFolderNotFoundError } from "@zuse/contracts";
 
-import { WorktreeService } from "../../worktree/services/worktree-service.ts";
+import { WorktreeService } from "@zuse/git/worktree-service";
 import {
   FileSearchService,
   type FileSearchHit,
